Integrate native additions
