--- android/app/src/main/kotlin/com/example/yumsg/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.yumsg
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()
